refactor(uhk-common): logging utilities to resolve circular dependencies
