+++ packages/uhk-common/src/index.ts
@@ -1,5 +1,5 @@
 export { Buffer } from 'buffer/';
-export * from './util';
-export * from './models';
-export * from './services';
 export * from './config-serializer';
+export * from './log';
+export * from './models';
+export * from './util';
