+++ packages/uhk-common/src/log/default-log-options.ts
@@ -1,0 +1,5 @@
+import { LogOptions } from '../models';
+
+export const DEFAULT_LOG_OPTIONS: LogOptions = {
+    misc: true
+};
