+++ packages/uhk-common/src/util/index.ts
@@ -1,12 +1,9 @@
 export { IpcEvents } from './ipcEvents';
-export * from './log';
 export * from './constants';
 export * from './create-md5-hash';
-export * from './get-log-options';
 export * from './get-md5-hash-from-file-name';
 export * from './helpers';
 export * from './is-equal-array';
-export * from './log-user-config-helper';
 export * from './sort-string-desc';
 export * from './user-configuration-history-helpers';
